Update FloatingProseContainer so not a container, rename to Highlight
