+++ src/components/Contentful/_contentful_new_design.scss
@@ -4,7 +4,7 @@
   @import 'components/Contentful/contentful_settings';
   @import 'components/Contentful/contentful_fonts';
   @import 'components/Contentful/contentful_hero';
-  @import 'components/Contentful/contentful_floating_prose_container';
+  @import 'components/Contentful/contentful_highlight';
 
   /** DEFAULT PAGE TEXT-STYLES */
